+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # BlogMore ChangeLog
 
+## v1.9.1
+
+**Released: 2026-03-10**
+
+- Fixed an error when using a relative path for the `output` configuration
+  option. Using a relative path such as `site` now works correctly.
+  ([#242](https://github.com/davep/blogmore/pull/242))
+
 ## v1.9.0
 
 **Released: 2026-03-09**